+++ packages/client/src/play/index.js
@@ -1,2 +1,3 @@
 export { createLivePlayClient } from './live.js'
+export * from './chrome.js'
 export * from './schema.js'